Feature/initial project

+++ src/main/java/br/com/priscyladepaula/avaliacaoekan/repository/BeneficiarioRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.priscyladepaula.avaliacaoekan.model.Beneficiario;
 
 @Repository
-public interface BeneficiarioRepository extends JpaRepository<Beneficiario, Integer>{
+public interface BeneficiarioRepository extends JpaRepository<Beneficiario, Long>{
     
 }
